requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
jinja2==3.1.2
python-dotenv==1.0.0
httpx==0.25.1
pytest==7.4.3
aiohttp==3.8.6
python-multipart==0.0.6
